Implemented Filter and Client service+controller
I implemented a generic filter that can be used across the program to be able to log the action taking place. I also implemented Client Service and Controller to be able to test this implementation. For future use, client must not use the list inside of the ClientService.cs file. I also implemented a REST file to test the implementation. I changed the model to be able to test the filters imlpementation. I will revert the changes on the model once we have the DB set up